+++ avrohugger-tools/src/test/compiler/output-scavro/avro/examples/baseball/Player.scala
@@ -30,7 +30,7 @@ object Player {
     override val avroClass: Class[JPlayer] = classOf[JPlayer]
     override val schema: Schema = JPlayer.getClassSchema()
     override val fromAvro: (JPlayer) => Player = {
-      (j: JPlayer) => Player(j.getNumber.toInt, j.getFirstName.toString, j.getLastName.toString, Array((j.getNicknames.asScala: _*)) map { x =>
+      (j: JPlayer) => Player(j.getNumber, j.getFirstName.toString, j.getLastName.toString, Array((j.getNicknames.asScala: _*)) map { x =>
         Nickname.metadata.fromAvro(x)
       })
     }
